chore(react): refactor presentation components
